clean up fid
