+++ examples/generative/conf/config_fid.yaml
@@ -1,0 +1,42 @@
+# Copyright (c) 2023, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+hydra:
+  job:
+    chdir: True
+  run:
+    dir: ./outputs/
+
+# Main options 
+mode: calc                # calc: calculate FID for a given set of images
+                          # ref: Calculate dataset reference statistics needed by 'calc'
+
+ # FID options
+image_path: ./images      # Path to the images
+ref_path: ./ref           # Dataset reference statistics
+num_expected: 50000       # Number of images to use
+seed: 0                   # Random seed for selecting the images
+batch: 64                 # Maximum batch size
+
+# Reference statistics options
+dataset_path: ./data      # Path to the dataset
+dest_path: ./dest.npz     # Destination .npz file
+batch: 64                 # Maximum batch size
+
+
+
+
+
+
+
